+++ modules/devtools/genrevibes_devtools/lib/src/dev_tools_hub_screen.dart
@@ -100,6 +100,7 @@ class StarterKitLabScreen extends StatelessWidget {
                   provider: host.ads!,
                   policy: host.adPolicy,
                   placements: host.adPlacements,
+                  switches: host.developerAdSwitches,
                 ),
         missing: 'genrevibes_ads',
       ),